move Age to remove circular imports, move AccordionItem up a level
