+++ containers/ecr-viewer/src/app/services/types.ts
@@ -1,0 +1,5 @@
+export interface Age {
+  years: number;
+  months: number;
+  days: number;
+}
