+++ containers/ecr-viewer/src/app/types.ts
